Pcp fixes (#11)
* Made a big start on fixing with the pcp plugin
Not even halfway done yet it seems

* Deleted premium functions
these functions where in the attachment.php and had errors/forbidden errors with the PCP plugin
All but 3 methods remain

* added comments mostly because of the size of the issue
Issue is that the nonces and validations on globals are not there these need to be implemented first before releasing the plugin

* Allergen update
In the allergen file most get variables are taken out
But while doing so the show allergen file needs rework to make certain functions work again
When reworking both files I am afraid that the whole table needs to be remade into a working proper PCP allowed class

* fixed a part of the filter nonce

* form nonces
All forms have been updated where neccesary with nonces
In the allergen_form the unused forms like the update and other premium features have been removed
Some database functions for the table have been removed as well either because it is a premium version or it was deemed unneccesary

* database cleanup and testing
Cleaning up last pcp error in the allergen db class

* db caching and direct call
When using db functions all calls towards our own custom tables will be labeled as a warning
To remove the warnings on caching and use of a direct call when using the plugin checker use the following comment:
// phpcs:ignore WordPress.DB.DirectDatabaseQuery

* DB files
all DB files in the DB directory have no more false positives on errors
All Warnings on direct call are now ignored since we use a custom DB scheme
Filtering also works again

* All db errors gone
This includes the warnings given by directly calling the db
Also updated the tables to make sure that there is no undefined error
Still need to check what the update_modification_detected code is and fix it
After that the readme file needs to be cleaned and updated
When all that is done the UML is in need of cleaning up

* Main file update
When checking with PCP the auto update functions and changelog functions are highly discouraged by wordpress
It gives all kinds of errors back when implemented so as for now until the release
The changelog and auto update functions are removed until further notice

* README and readme
Fixed the errors and warnings given by pcp
All updates and improvements still need to be added in the readme only after the md is updated after a review
md version can be placed in main folder before submitting to wordpress
When testing the plugin do not look at the UML as it will be updated later when all seems well and works

* deleted old check for the activator

+++ allergens-dietary/php/woocommerce/product_settings.php
@@ -32,7 +32,8 @@ class Allergens_Dietary_Product_Settings
 
 	public function __construct()
 	{
-		$this->_allergens = Allergens_Dietary_Allergen_Queries::getInstance()->getAllAllergens();
+		$this->_allergens = is_null(Allergens_Dietary_Allergen_Queries::getInstance()->getAllAllergens()) ? array() : 
+		Allergens_Dietary_Allergen_Queries::getInstance()->getAllAllergens();
 
 		add_filter('woocommerce_product_data_tabs', array($this, 'data_tab'));
 		add_action('woocommerce_product_data_panels', array($this, 'data_fields'));
@@ -77,6 +78,7 @@ class Allergens_Dietary_Product_Settings
 
 		?>
 		<div id="allergens_dietary_ictoria_product_data" class="panel woocommerce_options_panel">
+			<?php wp_nonce_field('allergen-product-action','allergen-product-nonce'); ?>
 		<h2><?php echo esc_html__( 'Select allergen(s) and/or dietary restrictions:', 'allergens-dietary' ); ?></h2>
 		<?php foreach ( $allergens as $allergen ) : ?>
 			<?php 
@@ -120,38 +122,46 @@ class Allergens_Dietary_Product_Settings
 	 */
 	public function save_product_options($post_id)
 	{
-		$allergensSelected = array();
-		$allergenNames = array();
-		unset($this->_attachedAllergens);
-		$this->_attachedAllergens = Allergens_Dietary_Allergy_Product_Queries::getInstance()->getAllergyProduct($post_id);
-
-		foreach ($this->_allergens as $allergen) {
-			if (isset($_POST[($this->replace_space_chars($allergen['allergy_name']) . '_allergens_dietary_ictoria')])) {
-				$allergensSelected[] = $allergen['allergy_name'];
+		if(isset($_POST['allergen-product-nonce']) && !empty($_POST['allergen-product-nonce']))
+		{
+			$nonce = sanitize_text_field(wp_unslash($_POST['allergen-product-nonce']));
+			if(!wp_verify_nonce($nonce, 'allergen-product-action')){
+				wp_die(esc_html(__('Something went wrong!','allergens-dietary')));
 			}
-		}
-		// Set old allergen name settings
-		foreach ($this->_attachedAllergens as $allergen) {
-			$allergenNames[] = $allergen['allergy_name'];
-		}
-		
-		$new_diff = array_diff($allergensSelected, $allergenNames);
-		$old_diff = array_diff($allergenNames, $allergensSelected);
-
-		// If the new doesnt contain allergens from the old one, Delete the old.
-		if ($old_diff) {
-			foreach ($old_diff as $allergen) {
-				Allergens_Dietary_Allergy_Product_Queries::getInstance()->deleteAllergyProduct($post_id, $allergen);
+			
+			$allergensSelected = array();
+			$allergenNames = array();
+			unset($this->_attachedAllergens);
+			$this->_attachedAllergens = Allergens_Dietary_Allergy_Product_Queries::getInstance()->getAllergyProduct($post_id);
+	
+			foreach ($this->_allergens as $allergen) {
+				if (isset($_POST[($this->replace_space_chars($allergen['allergy_name']) . '_allergens_dietary_ictoria')])) {
+					$allergensSelected[] = $allergen['allergy_name'];
+				}
 			}
-		}
-		// If the old doesn't contain allergens from the new one, Add the new.
-		if ($new_diff) {
-			foreach ($new_diff as $allergen) {
-				Allergens_Dietary_Allergy_Product_Queries::getInstance()->addAllergyProduct($post_id, $allergen);
+			// Set old allergen name settings
+			foreach ($this->_attachedAllergens as $allergen) {
+				$allergenNames[] = $allergen['allergy_name'];
 			}
+			
+			$new_diff = array_diff($allergensSelected, $allergenNames);
+			$old_diff = array_diff($allergenNames, $allergensSelected);
+	
+			// If the new doesnt contain allergens from the old one, Delete the old.
+			if ($old_diff) {
+				foreach ($old_diff as $allergen) {
+					Allergens_Dietary_Allergy_Product_Queries::getInstance()->deleteAllergyProduct($post_id, $allergen);
+				}
+			}
+			// If the old doesn't contain allergens from the new one, Add the new.
+			if ($new_diff) {
+				foreach ($new_diff as $allergen) {
+					Allergens_Dietary_Allergy_Product_Queries::getInstance()->addAllergyProduct($post_id, $allergen);
+				}
+			}
+	
+			return;
 		}
-
-		return;
 	}
 
 	private function replace_space_chars(string $allergens): string
